cleaned the code and fixed the bug with editing of todo item
