+++ homeworks/oleg-gorodenskyi/19-react-hooks/src/App.css
@@ -46,6 +46,7 @@ input::placeholder {
     font-style: normal;
     font-weight: 400;
     line-height: normal;
+    cursor: pointer;
 }
 
 .toDoItem {
